backend/requirements.txt
# FastAPI framework
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# HTTP clients
httpx==0.24.0
aiohttp==3.8.4

# Vector similarity search
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
nltk==3.8.1

# Web scraping
zyte-api==0.4.0
beautifulsoup4==4.12.0
playwright==1.36.0

# Utilities
pydantic==1.10.8
pydantic-settings==2.0.3